installer: download pack files straight to .Download/
Also make sure to re-use a previously downloaded file
before making a new HTTP request to retrieve the file.

+++ cmd/utils/utils.go
@@ -29,6 +29,10 @@ func DownloadFile(URL string) (string, error) {
 	parsedURL, _ := url.Parse(URL)
 	filePath := path.Join(CacheDir, path.Base(parsedURL.Path))
 	log.Debugf("Downloading %s to %s", URL, filePath)
+	if FileExists(filePath) {
+		log.Debugf("Download not required, using the one from cache")
+		return filePath, nil
+	}
 
 	out, err := os.Create(filePath)
 	if err != nil {
